# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdOps - Ad Account Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  provider: "provider",
  pitcher: "pitcher",
} as const;

export interface User {
  id: number;
  username: string;
  displayName: string;
  role: UserRole;
  /** @nullable */
  portalSlug?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  provider: "provider",
  pitcher: "pitcher",
} as const;

export interface CreateUserBody {
  username: string;
  displayName: string;
  password: string;
  role: CreateUserBodyRole;
  portalSlug: string;
}

export interface UpdateUserBody {
  displayName?: string;
  /** @nullable */
  password?: string | null;
  isActive?: boolean;
  /** @nullable */
  portalSlug?: string | null;
}

export type AccountPlatform =
  (typeof AccountPlatform)[keyof typeof AccountPlatform];

export const AccountPlatform = {
  FB: "FB",
  GG: "GG",
  TT: "TT",
  TW: "TW",
  OTHER: "OTHER",
} as const;

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus];

export const AccountStatus = {
  idle: "idle",
  active: "active",
  banned: "banned",
} as const;

export interface Account {
  id: number;
  platformAccountId: string;
  accountName: string;
  platform: AccountPlatform;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  /** @nullable */
  pitcherId?: number | null;
  /** @nullable */
  pitcherName?: string | null;
  status: AccountStatus;
  currentBalance: string;
  /** @nullable */
  theoreticalBalance?: string | null;
  /** @nullable */
  lastReportedAt?: string | null;
  createdAt: string;
}

export type CreateAccountBodyPlatform =
  (typeof CreateAccountBodyPlatform)[keyof typeof CreateAccountBodyPlatform];

export const CreateAccountBodyPlatform = {
  FB: "FB",
  GG: "GG",
  TT: "TT",
  TW: "TW",
  OTHER: "OTHER",
} as const;

export interface CreateAccountBody {
  platformAccountId: string;
  accountName: string;
  platform: CreateAccountBodyPlatform;
  initialBalance: string;
}

export interface UpdateAccountBody {
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface AssignAccountBody {
  /** @nullable */
  pitcherId: number | null;
}

export interface DailyStat {
  id: number;
  accountId: number;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  platformAccountId?: string | null;
  date: string;
  spendAmount: string;
  realBalance: string;
  pitcherId: number;
  /** @nullable */
  pitcherName?: string | null;
  hasAlert: boolean;
  createdAt: string;
}

export interface CreateDailyStatBody {
  accountId: number;
  date: string;
  spendAmount: string;
  realBalance: string;
}

export interface UpdateDailyStatBody {
  /** @nullable */
  spendAmount?: string | null;
  /** @nullable */
  realBalance?: string | null;
}

export type RechargeOrderStatus =
  (typeof RechargeOrderStatus)[keyof typeof RechargeOrderStatus];

export const RechargeOrderStatus = {
  pending: "pending",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface RechargeOrder {
  id: number;
  accountId: number;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  platformAccountId?: string | null;
  amount: string;
  status: RechargeOrderStatus;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  /** @nullable */
  pitcherId?: number | null;
  /** @nullable */
  pitcherName?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRechargeOrderBody {
  accountId: number;
  amount: string;
  /** @nullable */
  note?: string | null;
}

export type UpdateRechargeOrderBodyStatus =
  (typeof UpdateRechargeOrderBodyStatus)[keyof typeof UpdateRechargeOrderBodyStatus];

export const UpdateRechargeOrderBodyStatus = {
  completed: "completed",
  rejected: "rejected",
} as const;

export interface UpdateRechargeOrderBody {
  status: UpdateRechargeOrderBodyStatus;
  /** @nullable */
  note?: string | null;
}

export interface DashboardSummary {
  totalAccounts: number;
  idleAccounts: number;
  activeAccounts: number;
  bannedAccounts: number;
  totalProviders: number;
  totalPitchers: number;
  todayTotalSpend: string;
  pendingRechargeOrders: number;
  alertCount: number;
}

export interface SpendByProvider {
  providerId: number;
  providerName: string;
  todaySpend: string;
  totalSpend: string;
  accountCount: number;
}

export interface SpendByPitcher {
  pitcherId: number;
  pitcherName: string;
  todaySpend: string;
  totalSpend: string;
  accountCount: number;
}

export interface CrossReportRow {
  pitcherId: number;
  pitcherName: string;
  providerId: number;
  providerName: string;
  totalSpend: string;
  accountCount: number;
}

export interface BalanceAlert {
  accountId: number;
  accountName: string;
  platformAccountId: string;
  /** @nullable */
  pitcherId?: number | null;
  /** @nullable */
  pitcherName?: string | null;
  theoreticalBalance: string;
  reportedBalance: string;
  discrepancyPct: number;
  /** @nullable */
  lastReportedAt?: string | null;
}

export type ListUsersParams = {
  role?: ListUsersRole;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  provider: "provider",
  pitcher: "pitcher",
} as const;

export type ListAccountsParams = {
  status?: ListAccountsStatus;
  /**
   * @nullable
   */
  providerId?: number | null;
  /**
   * @nullable
   */
  pitcherId?: number | null;
};

export type ListAccountsStatus =
  (typeof ListAccountsStatus)[keyof typeof ListAccountsStatus];

export const ListAccountsStatus = {
  idle: "idle",
  active: "active",
  banned: "banned",
} as const;

export type ListDailyStatsParams = {
  /**
   * @nullable
   */
  accountId?: number | null;
  /**
   * @nullable
   */
  pitcherId?: number | null;
  /**
   * @nullable
   */
  dateFrom?: string | null;
  /**
   * @nullable
   */
  dateTo?: string | null;
};

export type ListRechargeOrdersParams = {
  status?: ListRechargeOrdersStatus;
  /**
   * @nullable
   */
  accountId?: number | null;
};

export type ListRechargeOrdersStatus =
  (typeof ListRechargeOrdersStatus)[keyof typeof ListRechargeOrdersStatus];

export const ListRechargeOrdersStatus = {
  pending: "pending",
  completed: "completed",
  rejected: "rejected",
} as const;

export type GetSpendByProviderParams = {
  /**
   * @nullable
   */
  dateFrom?: string | null;
  /**
   * @nullable
   */
  dateTo?: string | null;
};

export type GetSpendByPitcherParams = {
  /**
   * @nullable
   */
  dateFrom?: string | null;
  /**
   * @nullable
   */
  dateTo?: string | null;
};

export type GetCrossReportParams = {
  /**
   * @nullable
   */
  pitcherId?: number | null;
  /**
   * @nullable
   */
  providerId?: number | null;
};