# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdOps - Ad Account Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Users (Admin only) ───────────────────────────────────────────────────
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List ad accounts (role-filtered automatically)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [idle, active, banned]
        - in: query
          name: providerId
          schema:
            type: ["integer", "null"]
        - in: query
          name: pitcherId
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create/upload a new ad account (provider)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account (admin or provider)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}/assign:
    post:
      operationId: assignAccount
      tags: [accounts]
      summary: Assign an idle account to a pitcher (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAccountBody"
      responses:
        "200":
          description: Account assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  # ─── Daily Stats ──────────────────────────────────────────────────────────
  /daily-stats:
    get:
      operationId: listDailyStats
      tags: [daily-stats]
      summary: List daily spend reports (role-filtered)
      parameters:
        - in: query
          name: accountId
          schema:
            type: ["integer", "null"]
        - in: query
          name: pitcherId
          schema:
            type: ["integer", "null"]
        - in: query
          name: dateFrom
          schema:
            type: ["string", "null"]
        - in: query
          name: dateTo
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of daily stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStat"

    post:
      operationId: createDailyStat
      tags: [daily-stats]
      summary: Submit daily spend report (pitcher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyStatBody"
      responses:
        "201":
          description: Daily stat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyStat"
  /daily-stats/{id}:
    patch:
      operationId: updateDailyStat
      tags: [daily-stats]
      summary: Update a daily stat entry (pitcher)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyStatBody"
      responses:
        "200":
          description: Updated daily stat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyStat"

  # ─── Recharge Orders ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Auth ─────────────────────────────────────────────────────────────
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user

    # ─── Users ────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, provider, pitcher]
        portalSlug:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - role
        - isActive
        - createdAt
...
    Account:
      type: object
      properties:
        id:
          type: integer
        platformAccountId:
          type: string
        accountName:
          type: string
        platform:
          type: string
          enum: [FB, GG, TT, TW, OTHER]
        providerId:
          type: integer
        providerName:
          type: ["string", "null"]
        pitcherId:
          type: ["integer", "null"]
        pitcherName:
          type: ["string", "null"]
        status:
          type: string
          enum: [idle, active, banned]
        currentBalance:
          type: string
        theoreticalBalance:
          type: ["string", "null"]
        lastReportedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - platformAccountId
        - accountName
        - platform
        - providerId
        - status
        - currentBalance
        - createdAt
    CreateAccountBody:
      type: object
      properties:
        platformAccountId:
          type: string
        accountName:
          type: string
        platform:
          type: string
          enum: [FB, GG, TT, TW, OTHER]
        initialBalance:
          type: string
      required:
        - platformAccountId
        - accountName
        - platform
        - initialBalance
    UpdateAccountBody:
      type: object
      properties:
        accountName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    AssignAccountBody:
      type: object
      properties:
        pitcherId:
          type: ["integer", "null"]
      required:
        - pitcherId

    # ─── Daily Stats ──────────────────────────────────────────────────────
    DailyStat:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        accountName:
          type: ["string", "null"]
        platformAccountId:
          type: ["string", "null"]
        date:
          type: string
        spendAmount:
          type: string
        realBalance:
          type: string
        pitcherId:
          type: integer
        pitcherName:
          type: ["string", "null"]
        hasAlert:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - accountId
        - date
        - spendAmount
        - realBalance
        - pitcherId
        - hasAlert
        - createdAt
    CreateDailyStatBody:
      type: object
      properties:
        accountId:
          type: integer
        date:
          type: string
        spendAmount:
          type: string
        realBalance:
          type: string
      required:
        - accountId
        - date
        - spendAmount
        - realBalance
    UpdateDailyStatBody:
      type: object
      properties:
        spendAmount:
          type: ["string", "null"]
        realBalance:
          type: ["string", "null"]

    # ─── Recharge Orders ──────────────────────────────────────────────────
...